# Api-Backend-LocSystem/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocSystem API",
        "description": "Documentação da API do sistema LocSystem",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/operators": {
            "get": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Lista os localizadores cadastrados com suporte a filtros administrativos.",
                "operationId": "cb3534cd600d6fc1baf21ec79dcd4ae1",
                "responses": {
                    "200": {
                        "description": "Localizadores retornados com sucesso no campo operators."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem consultar localizadores."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao listar os localizadores."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Cadastra um novo localizador com dados de acesso e assinatura.",
                "operationId": "c507826c26152f3b428b290a867ef869",
                "responses": {
                    "201": {
                        "description": "Localizador cadastrado com sucesso e retornado no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem cadastrar localizadores."
                    },
                    "409": {
                        "description": "E-mail já utilizado ou plano obrigatório não informado para localizador mensalista."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados para cadastro do localizador."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao cadastrar o localizador."
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 64
                                    },
                                    "v_document": {
                                        "type": "string"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 11
                                    },
                                    "i_user_limit": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "b_is_courtesy": {
                                        "type": "boolean"
                                    },
                                    "i_pricing_plan_id": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_password",
                                    "v_document",
                                    "v_phone",
                                    "i_pricing_plan_id"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/operators/{id}": {
            "get": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Retorna os detalhes de um localizador específico.",
                "operationId": "8d2a147b41639f46bc4d711710680b9a",
                "responses": {
                    "200": {
                        "description": "Localizador retornado com sucesso no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado ou localizador não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao consultar o localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Atualiza os dados cadastrais e comerciais de um localizador.",
                "operationId": "ea822e842f49934e88fb914174267b0d",
                "responses": {
                    "200": {
                        "description": "Localizador atualizado com sucesso e retornado no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado ou localizador não encontrado."
                    },
                    "409": {
                        "description": "Conflito de e-mail, limite de prepostos inválido ou plano obrigatório não informado."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados para atualização do localizador."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "v_document": {
                                        "type": "string"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 11
                                    },
                                    "i_user_limit": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "b_is_courtesy": {
                                        "type": "boolean"
                                    },
                                    "i_pricing_plan_id": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_document",
                                    "v_phone",
                                    "i_pricing_plan_id"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Remove um localizador e seus vínculos administrativos relacionados.",
                "operationId": "2ff042c830fd9c75d1a29f83073dd650",
                "responses": {
                    "200": {
                        "description": "Localizador removido com sucesso e confirmação retornada no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado ou localizador não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao remover o localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/operators/{id}/approval": {
            "patch": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Aprova um localizador para acesso ao sistema.",
                "operationId": "f03da2fa4d37996a1ed12b79c073b597",
                "responses": {
                    "200": {
                        "description": "Localizador aprovado com sucesso e dados atualizados retornados no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado ou localizador não encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao aprovar o localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/operators/{id}/status": {
            "patch": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Bloqueia ou desbloqueia um localizador conforme o status informado.",
                "operationId": "7b9ba40c0b7ece3d0be3ce8d758985d4",
                "responses": {
                    "200": {
                        "description": "Status do localizador atualizado com sucesso e dados retornados no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado ou localizador não encontrado."
                    },
                    "409": {
                        "description": "O desbloqueio não pode ser realizado porque a assinatura do localizador não está ativa."
                    },
                    "422": {
                        "description": "Erro de validação no payload de alteração de status."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o status do localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 1000
                                    }
                                },
                                "required": [
                                    "isActive"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/operators/{id}/password": {
            "patch": {
                "tags": [
                    "Localizadores"
                ],
                "summary": "Define uma nova senha para o localizador informado.",
                "operationId": "ea08aa315358c0e76aa1a96c373093b8",
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso e confirmação retornada no campo operator."
                    },
                    "401": {
                        "description": "Acesso negado, localizador não encontrado ou conta de autenticação inexistente."
                    },
                    "422": {
                        "description": "Erro de validação da nova senha informada."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao alterar a senha do localizador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 64
                                    }
                                },
                                "required": [
                                    "v_password"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}